Finish up moving to jdbc

+++ src/main/java/com/financeplanner/datasource/mapper/UserMapper.java
@@ -13,7 +13,7 @@ public class UserMapper implements RowMapper<User> {
     public User mapRow(ResultSet resultSet, int i) throws SQLException {
         User user = new User();
 
-        user.setId(resultSet.getInt("id"));
+        user.setId(resultSet.getLong("id"));
         user.setEmail(resultSet.getString("email"));
         user.setName(resultSet.getString("name"));
         user.setPassword(resultSet.getString("password"));
